Update README.md
Hospital Playlist is not solely a love story, but it makes viewers swoon and intrigued. It's light but never boring. It's a musical but relatable and realistic.

Oh, My Venus is a beautiful drama with a great story and great chemistry between the main leads. There is a lot of romance among the main leads and the viewers can see that love doesn't have anything to do with body shape, size, and health.